+++ packages/expo-router/src/react-navigation/native-stack/index.tsx
@@ -12,11 +12,13 @@ export { NativeStackView } from './views/NativeStackView';
  * Hooks
  */
 export { useAnimatedHeaderHeight } from './utils/useAnimatedHeaderHeight';
+export { makePopAction } from './utils/makePopAction';
 
 /**
  * Types
  */
 export type {
+  NativeStackEmit,
   NativeStackHeaderBackProps,
   NativeStackHeaderItem,
   NativeStackHeaderItemButton,
@@ -32,10 +34,12 @@ export type {
   NativeStackHeaderNativeProps,
   NativeStackNativeProps,
   NativeStackNavigationEventMap,
+  NativeStackDescriptorMap,
   NativeStackNavigationOptions,
   NativeStackNavigationProp,
   NativeStackNavigatorProps,
   NativeStackOptionsArgs,
   NativeStackScreenNativeProps,
   NativeStackScreenProps,
+  NativeStackViewEmit,
 } from './types';